Extract EncryptionService from MessageDB
